+++ kernel/src/main/java/com/itextpdf/kernel/font/DocFontEncoding.java
@@ -117,7 +117,7 @@ class DocFontEncoding extends FontEncoding {
                     currentNumber = ((PdfNumber) obj).intValue();
                 } else {
                     String glyphName = ((PdfName) obj).getValue();
-                    int unicode = (int) AdobeGlyphList.nameToUnicode(glyphName);
+                    int unicode = AdobeGlyphList.nameToUnicode(glyphName);
                     if (unicode != -1) {
                         fontEncoding.codeToUnicode[currentNumber] = (int) unicode;
                         fontEncoding.unicodeToCode.put((int) unicode, currentNumber);
